+++ pyobjc-framework-GameKit/PyObjCTest/test_gkcloudplayer.py
@@ -1,0 +1,10 @@
+from PyObjCTools.TestSupport import *
+
+import GameKit
+
+class TestGKCloudPlayer (TestCase):
+    def testMethods(self):
+        self.assertArgIsBlock(GameKit.GKCloudPlayer.getCurrentSignedInPlayerForContainer_completionHandler_, 1, b'v@@')
+
+if __name__ == "__main__":
+    main()