+++ pyobjc-framework-GameKit/PyObjCTest/test_gkgamesessioneventlistener.py
@@ -1,0 +1,12 @@
+from PyObjCTools.TestSupport import *
+
+import objc
+import GameKit
+
+class TestGKGameSessionEventListener (TestCase):
+    def testProtocols(self):
+        objc.protocolNamed('GKGameSessionEventListener')
+
+
+if __name__ == "__main__":
+    main()