+++ pyobjc-framework-GameKit/PyObjCTest/test_gkleaderboardviewcontroller.py
@@ -1,0 +1,12 @@
+from PyObjCTools.TestSupport import *
+
+import objc
+import GameKit
+
+class TestGKLeaderboardViewController (TestCase):
+    def testProtocols(self):
+        objc.protocolNamed('GKLeaderboardViewControllerDelegate')
+
+
+if __name__ == "__main__":
+    main()