+++ pyobjc-framework-GameKit/PyObjCTest/test_gksavedgamelistener.py
@@ -1,0 +1,10 @@
+from PyObjCTools.TestSupport import *
+
+import GameKit
+
+class TestGKSavedGameListener (TestCase):
+    def testProtocols(self):
+        objc.protocolNamed('GKSavedGameListener')
+
+if __name__ == "__main__":
+    main()